remove unneeded mock viewmodel
